Adjust project description and add license info
This change is in response to the introduction of AWS Batch,
which makes it easier to manage ECS tasks for data analysis,
mostly obviating the need for an advanced scheduler.

We have also changed the scope of the repo to implement
an entire API for data management and job scheduling
in a single location, both for conciseness and consistency.

We also added license info (or lack thereof) to ensure that
this package is not accidentally published with unintended consequences.
A decision on how to properly share this work will be made later.