Implement first working version of the init script for Batch-managed instances
This feature enables automatic creation, attachment, formatting
and mounting of EBS volumes to Batch-managed (or any other) EC2 instances,
using a Lambda function triggered on EC2 instance 'running' events.

The function makes appropriate sanity checks to ensure
that volume attachment completes successfully, or deletes
them after an unsuccessful attempt.

It also calculates an optimal 'gp2' volume configuration
that matches EBS network performance for that instance type.

After attachment, it fetches a pre-defined SSH key from S3
and attempts to run a remote shell script to format the drive
on the instance, using either the single attached volume device,
or creating a RAID0 array prior to that. Finally, it mounts
the new drive on the instance.

The Lambda function is heavily unit-tested (albeit some edge cases
may still remain) with 100% coverage, and hand-verified
to actually work with Batch. We still need to add
integration tests for completeness.