Initial definitions for the Swagger framework
This change sets up the Swagger framework
to develop and test our APIs in Swagger Editor.

Changes are immediately saved and unit-tested.
In addition, we concatenate SAM and Swagger templates
to enable use of CloudFormation references in the latter.

We have also defined an authorizer in the Swagger template,
but haven't created an appropriate function in SAM yet,
so the CloudFormation update will fail.