Remove x-api-key from Swagger and SAM
It turns out that this key is ineffective, because
the Authorizer doesn't use it in the current implementation
of API Gateway by AWS.

In addition, it unfortunately confuses Swagger Codegen
because both member_token and api_key map to the same
apiKey variable on the client.

We may reintroduce it in the future when these issues are resolved.