Add ability to reuse a hosted zone for CloudFormation
Previously, if the backend stack was deleted, the hosted zone
would be gone. Next time the stack was created, the new zone
would contain different DNS servers from the ones
registered for the domain, effectively making this zone
incapable of serving DNS requests.

Now, the stack creator may choose to reuse an existing zone,
thus minimizing the chance of mismatch with domain registry.

However, even if a mismatch happens, it is still possible to
re-assign DNS servers manually (e.g. through AWS Console).
We plan to include instructions on how to do that in
an upcoming README.