Add explicit dependency on API stage for API base path mapping
Unfortunately, API base mapping resource does not appear to wait
for AWS::Serverless:Api resource to be fully deployed
before it tries to reference its stage.

We have now included its explicit dependency on the stage name.
This naming convention is not well-documented, so we may
soon replace it with a 'regular' AWS::ApiGateway::RestApi,
along with Deployment and Stage resources instead
(that will also give us more flexibility in their configuration).