Use LogicalResourceId instead of RequestId as CallerReference for OAI in SAM
This fix addresses the apparent requirement that CallerReference
must be the same in UpdateCloudFrontOriginAccessIdentity requests.
That contradicts the official docs, as request ID
is now *not* unique to that request.

However, if we do specify the same CallerReference,
then requests to UpdateCloudFrontOriginAccessIdentity
succeed, even if Comment (and thus CloudFrontOriginAccessIdentityConfig)
changes for the same CallerReference and ETag.