Wait for changesets in backend packaging project
This change addresses the issue when creation
of changesets is triggered successfully, but
the completion of creation would fail during deployment
due to some problems that were not caught
upon initial validation of the template.

We now wait for the changesets to be created
completely, and only then proceed to the next step.
This approach also enables us to retry
the packaging job if we manage to fix the
underlying issue ad hoc.