Use 'jsonwebtoken' library to parse JWT
We use Auth0's (yet cross-platform) library
to parse and verify JWT token. Previously,
we used 'jsrsasign' to do the same (at the cost
of reduced size), but that library is quirkier
and less popular nowadays.

Since we've decided that size of included libraries
no longer matters (up to a limit), we can now use
industry-standard tools without reinventing the wheel,
especially when it comes to security.