Share web signing key pair between 'beta' and 'release' stacks
Previously, we created separate key pairs for the 2 stacks,
but queried the first key pair ID in GenerateSignedCookies.

As a result, GenerateSignedCookies only worked for the 'beta' stack.
We now share the same key between the stacks.

While it creates an additional dependency between them,
it simplifies infrastructure setup as now we only need to import
one public key into security credentials.

We should caution stack admins not to delete the 'beta' stack
if they want to keep the 'release' distribution functional.

So we should consider 'beta' to be more of a 'preview'
that also sets up some important resources for 'release'.