Implement apig.validate() for request validation against the API spec
We implement a comprehensive validator that enumerates Swagger
parameter models for a particular resource and method,
and then validates each model against an appropriate part
of the Lambda proxy request.

Each part of request - path and query parameter(s), headers, and body -
is matched to a corresponding Swagger schema and validated
using a general-purpose JSON schema validator (Ajv).
This library was chosen for its popularity and high performance,
leading to a very minimal overhead (on the order of 10s of ms)
for the initial compilation of all schemas (on "cold" start),
as well as sub-ms overhead for the actual validation.
At higher request rates, Lambda instances can be reused,
and so this overhead becomes negligible.

apig.validate() is implemented through promises, and also
returns a full list of validation errors when appropriate.