Change CompressionRequest to 'string' type in Swagger
Previously, this request was of type 'array' but coarsed to
a CSV array by the spec. Unforunately, this schema is not
currently supported by Ajv out of the box. In addition,
real Accept-Encoding values may be set by the browser to more
complicated values, e.g. 'deflate, gzip;q=1.0, *;q=0.5'.
Finally, our implementation only checks for presense of
keywords such as 'deflate', 'gzip', or '*'.

As such, we've decided to define this request as 'string'
matched against the pattern of these keywords.