Tweak output of apig.respondWithError()
We now return a generic server error if error code is 500 or undefined.
In addition, we log the error prior to that.

We also remove 'errors' property from the final output if it is undefined.

Finally, we tweak ApiError class to allow seamless assignment of
Error to ApiError, which is helpful for responding with bare-bones Errors.