Add validation to web.generateSignedCookies()
Since we check for Accept-Encoding header in respond(),
we have to either make an exception for generateSignedCookies(),
or accept and validate this header for that request as well.

Here we opt for the second approach.
Additionally, it makes overall sense to run validate()
as the first check in all of our user-facing methods,
irregardless if they expect any arguments or not.

As such, we now also simplify and extend the check for
'event.requestContext.authorizer.expiresAt' to be done via a nested Promise.