Convert hit fields to single strings when possible in search.query()
It turns out that CloudSearchDomain.search() returns hit fields
as arrays of strings.

Accordingly, we need to convert them to single values, unless
they are comprised of more than one element.

In the future, we may need to treat this situation more carefully,
because some results may be legitimate single-element arrays,
or include non-string types (e.g. numbers).

For now, all of our results (as returned to the user) are defined
as singular strings.

As such, the current approach should work fine in the meantime.