Disable highlighting on 'description' in CloudSearch
Highlighting unnecessarily bloats off the index (~200%) and reduces
performance (at least, according to an official presentation from AWS).

Moreover, we don't want to modify the results in any way.

Instead, we will highlight matches in the client code,
as a bonus for graphical clients.

As a side benefit, that also simplifies formatting on the fly.