Remove Deny NotPrincipal policy for encryption key
We no longer need to explicitly deny access to anyone not
on the list of key policies.

But more importantly, assumed-role rules may no longer work
with some services, such as CodeBuild, as their session role names
are now generated dynamically by AWS, and NotPrincipal doesn't
support wildcards.

As such, we remove Deny NotPrincipal, even if it slightly widens
the scope of permissions for GenerateDataKey and Decrypt.
We still restrict it to particular roles that can only be
assumed by select services (Lambda and CodeBuild).
But we no longer restrict it to particular Lambda function or
CodeBuild project names (in the latter case it is no longer possible).