Convert incoming request body from base64 when needed
This conversion is necessary when we specify '*/*' as binary media type,
because API Gateway + Lambda proxy then treat any request as binary,
automatically encoding it to base64.

In turn, we decode requests automatically if isBase64Encoded flag is set.
That will help us to deal with both situations (encoded or not) going forward.