Remove conditions for DatasetsTable in SAM
In case we need to re-create the stack from an existing table,
we will provide a different mechanism to restore the data
(e.g. using Kinesis stream or DataPipeline export/import).

I.e. we will always create and use a new table,
optionally copying the data from an old one.