Implement PUT /datasets/{dataset_id}/storage for uploading -> available
We implement the first part of the dataset storage type setting,
where the user submits a request to make a dataset 'available'
once they finish uploading.

Currently, this method returns a 501 "Not Implemented" error
if they try to set storage type to 'archived', or if they try to 'restore'
from archive (in case the 'archived' status was somehow set prior to that).

We also increase the timeout to 30 sec and memory to 1GB.
That should provide enough time and processing power
to list all of the files that have been uploaded under that dataset's prefix,
and to apply project_id tag to each one (which is done in parallel).