Send only the latest DynamoDB stream records to CloudSearch
Previously, we didn't handle the situation when
a DynamoDB stream batches multiple updates to the same item
into a single record sequence.

Because each sequence is time-ordered, we can 'squeeze'
sequential updates for each item into a single record,
which will be the latest record in a batch.

This approach not only enhances performance, but in fact,
fixes incorrect processing of multiple updates to the same item,
because CloudSearch doesn't report an error when multiple
documents with the same id are uploaded in the same request.
Instead, it silently updates the item in an undocumented order.

Here, we fix this issue by turning the array of records into a Map,
which effectively stores only the latest version of each item.

We also improve the overall handling of item IDs by extracting
them from Keys.id for all event types.