Fix handling of output for deleted compute environments
It turns out that Batch reports these as 'null' instead of
a full definition with 'DELETED' status.

As such, we output the result of DescribeComputeEnvironments
to a special compartment, and compare it to 'null'
to make sure the environment is deleted, while preserving
the initial ResourceProperties (which may be used later to
re-create an environment).