Replace job queue inside state machine when the queue changes
Previously, we tried to use CloudFormation's official mechanism
for replacing resources. However, it didn't behave as we assumed
(i.e. a new 'Create' request did not get triggered, instead
a 'Delete' request was done).

Now, we explicitly check if jobQueueName has changed, then
remove and create a new queue in the same workflow if it did
(i.e. we mirror what we did for compute environments).