Add On-Demand compute environment
We create a mirror of SpotComputeEnvironment,
as a backup for situations where e.g. spot bid is too low.

Our queue will now depend on these two environments,
in order from Spot to On-Demand.

We also add an explicit dependency of the On-Demand environment
on Spot, which will get important during incompatible updates
(only one environment could be removed from the queue at a time,
because Batch queues disallow empty list of environments).