Ensure that compute environments are dissociated from the job queue prior to deletion
Before deletion of a compute environment, we now check
that its job queue has been deleted (e.g. upon deletion of the stack),
otherwise we explicitly remove the compute environment from the queue.

This dissociation from the queue is imposed by Batch.

However, because Batch doesn't allow us to dequeue the last remaining
compute environment, this process can only be done on one environment at a time
(so that another environment stays in the queue during an update,
unless we're removing the stack altogether).