Calculate md5 of the app/ directory based on its contents
Previously, we've calculated sha256 of the zip file,
which implicitly meant that we also hashed file metadata
(e.g. creation dates).

Since we only need to detect if the file contents has changed,
we can calculate a checksum recursively for the app/ folder.

Further, we can use md5 to speed this up.