Move job watching into ExecuteAnalysisStateMachine
We notice that setting up rules complicates logic of the backend,
since now we'd have to delete those rules when analysis ends
(while keeping in mind the limit on the number of rules).

Additionally, we now keep track of the old and new job statuses directly
inside the state machine, which makes it easier to avoid
unnecessary actions (e.g. calling DynamoDB) when they don't change.