Implement user-initiated cancellation of analyses
We implement DELETE /analyses/{analysis_id}/execution,
which enables user-initiated cancellation of an analysis.

This method only updates the dataset status to 'canceling'.
Then ExecuteAnalysisStateMachine picks up that change
the next time it detects a job status update.

If cancellation needs to happen more immediately,
we'll have to implement it differently (or check
the dataset status on every loop).