Implement job termination upon user request
Previously, we've terminated jobs indirectly by updating
the status of analysis to 'canceling'.

Now, we terminate jobs directly and then rely on
ExecuteAnalysisStateMachine for status updates and cleanup.

In contrast to the old implementation, we no longer
have to wait till a job has terminated
due to a "natural" success or failure.

Instead, jobs are terminated pre-emptively, and
the cleanup will happen in the next loop.