Explicitly update the initial status of an analysis in ExecuteAnalysisStateMachine
Previously, we waited for the analysis to change the status of one of its jobs.
Now, we set this status before the first loop such that the Analyses table
is updated sooner. This provides faster feedback to the user, and enables
them to cancel the analysis without waiting for the first job to be running.