Update analysis status in analyses.cancelExecution()
Previously, we used dynamodb.get() to only fetch the list of
job IDs to terminate. Now, we also conditionally update analysis status
to 'canceling' to indicate cancellation action to the user,
and to make sure that it can only occur when the current status
is compatible.

As a side benefit, we virtually eliminate the possibility that
job_ids will be naught or stale.