Always create a new search domain when re-creating the infrastructure stack
This is to ensure consistentcy with DynamoDB.
The database serves as the authoritative source of data,
and the search index should be rebuild from it.

When the administrator re-creates the stack, they will need
to import all of the data from the old tables to the new
ones, which will automatically trigger consistent indexing.