Add a Data stack with TagDataStateMachine
This state machine will be used to tag data on S3.

It calculates a hash of all object keys and their ETag hashes.

It then re-caculates this hash after a delay (currently 1 minute).

If the hash hasn't changed, it considers that the list of objects
is complete (i.e. they fully propagated in S3 due to eventual
consistency), and then tags associated objects.

This approach provides only a probabilistic guarantee that
the list of keys is stable, which should be true most of the time.