Exclude 'name' attribute from the list of required attribute for a user.
We only require it when creating a user through the API Gateway.

Even if a user was created without a name, their name would be
displayed as empty and should not break the UI client
(it may only degrade the experience).

Making this attribute optional simplifies requirements on
the external identity providers, which now only
need to supply the email.