Set email to be auto-verified on change for UserPool
It turns out that the purpose of an 'auto-verified' flag is to
automatically *send* a verification message to the user.

So in our case, if they change an email, a message with
the verification link will be automatically send to them.

TODO: We should check that the email is verified
when linking and authorizing a user identity.