Don't link to an internal Cognito user whose email is not verified
This fix prevents and exploit where an attacker could
"lure" the victim to sign up with their external provider
and be linked to the attacker's internal user identity
(using the same but unverified email), e.g. to fool
the victim into creating resources or storing information
that could then be accessed by the attacker.