Check if external user email is verified before trying to link it
Even though not all providers supply the email_verified attribute,
for those that do, we can check its value and refuse signup
if it is set to 'false'.

Currently, Google provides this attribute, while Facebook
describes that 'email' will be supplied only if it is valid
(and thus verified, we presume).