Implement UserPool as a custom resource
We now switch the user pool to be a custom resource,
because AWS::Cognito::UserPool doesn't support
UsernameAttributes and VerificationMessageTemplate.

UsernameAttributes is an immutable property
that is superior to aliases when it comes to email
and its verification. It also sets 'username'
to the value of 'sub' attribute for native users.

VerificationMessageTemplate enables verification links
instead of codes.

In addition, our custom resource prevents deletion of
the underlying user pool, and ignores updates to immutable
properties (PoolName, Schema, UsernameAttributes, AliasAttributes).