Define and implement methods to update and verify user attributes
We enable the user to update their own email and name,
and verify the new email with a temporary code.

This functionality is achieved through Cognito APIs that
require user's access token passed from the authorizer.

We also automatically send verification code even
if the value of the email didn't change.
That takes care of expired codes, where the user
is able to trigger a new verification by simply
applying the same update again.